requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
python-dotenv==1.0.1
supabase==1.1.1
PyJWT==2.9.0
APScheduler==3.11.1
reportlab==4.0.5
qrcode==7.4.2
Pillow==9.5.0
requests==2.32.0